requirements.txt
flask
flask-login
langchain>=0.0.300
chromadb
sentence-transformers
transformers
llama-cpp-python
python-docx
python-pptx
pypdf
tqdm
werkzeug
gunicorn
unstructured
unstructured-inference
langchain-embedding==0.0.1
flask-login
docx2txt

